==> 02_database_creation.ipynb <==
# Chroma Database Generation

# Library

from langchain_community.vectorstores import Chroma
from langchain_text_splitters import RecursiveCharacterTextSplitter
import os
import json
import uuid
import pickle as pkl
from langchain_huggingface import HuggingFaceEmbeddings
from langchain.chat_models import init_chat_model
import helper_functions as hf

## Config

with open('api_google.txt') as f:
    
    api_key = json.load(f)

if not os.environ.get("GOOGLE_API_KEY"):
  os.environ["GOOGLE_API_KEY"] = api_key['key']

## LLM model

llm = init_chat_model("gemini-2.5-flash", model_provider="google_genai",temperature=0,max_output_tokens=1024)

## Load data

The information for each article is structured in JSON format.

- `info_articles_main`: JSON file containing scientific papers from our research group.
- `info_articles_ref_final`: JSON file containing scientific papers referenced by the papers from our research group.

with open("info_articles_main.pkl","rb") as f:
    info_articles_main = pkl.load(f)
with open("info_articles_ref_final.pkl","rb") as f:
    info_articles_ref = pkl.load(f)

## Database

### Split the text

We use RecursiveCharacterTextSplitter. It splits text by recursively look at characters. Recursively tries to split by different characters to find one that works.

splitter = RecursiveCharacterTextSplitter(
    chunk_size=1000,  # We selected this threshold according to the performance of the model.
    chunk_overlap=100,  # 1000/10
    separators=["\n\n", "\n", ".", "!", "?", " "]  # smart splitting
)

We merge both JSON files into `info_articles_final`

info_articles_final = info_articles_main + info_articles_ref
f'We have a total of {len(info_articles_final)} scientific articles'

info_splitted = hf.chunk_makers(info_articles_final, splitter=splitter)

info_splitted[10]

## Embedding

We chose this embedding according to leaderboard of HuggingFace and the results of the performance of the model.

embedding_function2 = HuggingFaceEmbeddings(model_name="intfloat/e5-small-v2")

## Adding texts to ChromaDB

# Prepare documents, metadata, and IDs
texts = [chunk["content"] for chunk in info_splitted]
metadatas = [{"parent": chunk["parent"], "chunk_index": chunk["chunk_index"],"DOI": chunk["DOI"], "Reference": chunk["Reference"]} for chunk in info_splitted]
ids = [str(uuid.uuid1()) for _ in metadatas]

db = Chroma.from_texts(
    texts=texts,
    embedding=embedding_function2,
    metadatas=metadatas,
    ids=ids,
    collection_name="ReproRAG",
    persist_directory="./chromaRepro" # Persist directory to be read in the app.py (streamlit)
)

We check that the search works properly

db.similarity_search("Is there a signature to predict endometrial disruption?", 10) # Cosine by default
